trades

+++ api/src/routes/user.js
@@ -4,15 +4,14 @@ const { userSchema } = require('./validation')
 // user existence verified in authorization middleware
 
 async function update(req, res) {
-  const patch = req.body
-  const { error } = userSchema.validate(patch)
+  const { error } = userSchema.validate(req.body)
 
   if (error) {
     return res.status(400).send(error.details)
   }
 
   try {
-    const result = await User.update(patch, {
+    const result = await User.update(req.body, {
       where: { id: req.user.id },
       returning: true,
     })
@@ -31,6 +30,14 @@ async function remove(req, res) {
       where: { id: req.user.id },
       attributes: { exclude: ['password'] },
     })
+    // will this return both offeror and offeree?
+    const openTrades = await user.getOpenTrades()
+
+    // do not allow users with open trades to remove their account
+    if (openTrades.length > 0) {
+      return res.status(400).send('cannot deactivate account with open trades')
+    }
+
     await User.destroy({ where: { id: req.user.id } })
     res.send(user)
   } catch (error) {
